feat(606): Workaround fix health check implementation on minio (#150)
(cherry picked from commit b0411e6146231750a99e34f382b72634bb4c91c9)
